+++ gulp/all-tests.js
@@ -1,0 +1,3 @@
+var gulp = require('gulp');
+
+gulp.task('all-tests', ['unit-tests', 'integration-tests']);
